Added param for append or truncate-insert mode

+++ DataPipelines_WithAirflow/airflow/plugins/operators/load_dimension.py
@@ -18,18 +18,29 @@ class LoadDimensionOperator(BaseOperator):
                  redshift_conn_id="",
                  table="",
                  load_sql_stmt="",
+                 insert_mode="append",
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
         self.load_sql_stmt = load_sql_stmt
+        self.insert_mode = insert_mode
         
+                
     def execute(self, context):
         '''On run, delete dim table and reload data from insert SQL query and table name passed as an argument '''
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
-        redshift.run("DELETE FROM {}".format(self.table))
-        self.log.info("Deleted table : {}".format(self.table))
+        
+        if self.insert_mode == "truncate_insert":
+            self.log.info("Insert mode is truncate_insert Deleting table : {}".format(self.table))
+            redshift.run("DELETE FROM {}".format(self.table))
+            self.log.info("Deleted table : {}".format(self.table))
+        else:
+            self.log.info("Insert_mode = append. Appending data to Redshift  table: {} ..."\
+                            .format(self.table))
+            
+        
         self.log.info(f"Loading dimension table {self.table} in Redshift")
         formatted_sql = LoadDimensionOperator.insert_sql.format(
             self.table,
